+++ src/pages/Textareas.tsx
@@ -8,6 +8,89 @@ function Textareas() {
       <h3>Textareas</h3>
       <div className="lui-grid lui-grid-cols-3 lui-items-center lui-justify-center lui-gap-2 lui-text-center">
         <Textarea
+          inputStyle="underline"
+          label="Textarea"
+          placeholder="Type your message here."
+          regex={{ type: "number" }}
+        />
+        <Textarea
+          inputStyle="underline"
+          placeholder="Type your message here."
+          label="Label w/Tooltip"
+          tooltip={
+            <Tooltip
+              trigger={
+                <Icon
+                  name="question-filled"
+                  className="lui-text-ocean-primary-10"
+                />
+              }
+              side="top"
+              title="Tooltip Title"
+            >
+              <p>
+                Lorem ipsum dolor sit amet, consectetuer adipiscing elit. Aenean
+                commodo ligula eget dolor.
+              </p>
+            </Tooltip>
+          }
+        />
+        <Textarea
+          inputStyle="underline"
+          label="Textarea"
+          placeholder="Type your message here."
+          disabled
+        />
+      </div>
+      <div className="lui-grid lui-grid-cols-3 lui-items-center lui-justify-center lui-gap-2 lui-text-center">
+        <Textarea
+          inputStyle="underline"
+          label="Textarea"
+          placeholder="Type your message here."
+          alignment="horizontal"
+        />
+        <Textarea
+          inputStyle="underline"
+          placeholder="Type your message here."
+          alignment="horizontal"
+          label="Label w/Tooltip"
+          tooltip={
+            <Tooltip
+              trigger={
+                <Icon
+                  name="question-filled"
+                  className="lui-text-ocean-primary-10"
+                />
+              }
+              side="top"
+              title="Tooltip Title"
+            >
+              <p>
+                Lorem ipsum dolor sit amet, consectetuer adipiscing elit. Aenean
+                commodo ligula eget dolor.
+              </p>
+            </Tooltip>
+          }
+        />
+      </div>
+      <div className="lui-grid lui-grid-cols-3 lui-items-center lui-justify-center lui-gap-2 lui-text-center">
+        <Textarea
+          inputStyle="underline"
+          label="Textarea"
+          placeholder="Type your message here."
+          helperText="Helper Text"
+        />
+        <Textarea
+          inputStyle="underline"
+          label="Textarea"
+          placeholder="Type your message here."
+          errorMessage="Error message"
+        />
+      </div>
+
+      <div className="lui-grid lui-grid-cols-3 lui-items-center lui-justify-center lui-gap-2 lui-text-center">
+        <Textarea
+          id="asd"
           label="Textarea"
           placeholder="Type your message here."
           regex={{ type: "number" }}